# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Giang Minh Viet Apis",
    "version": "1.0.0",
    "title": "Giang Minh Viet Apis",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/product/": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productsPerPage",
            "in": "query",
            "description": "Max products can be display in one page",
            "required": false
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Page number",
            "required": false
          },
          {
            "name": "orderType",
            "in": "query",
            "description": "Order type to display: 1-oldest 2-newest ",
            "required": false
          },
          {
            "name": "search",
            "in": "query",
            "description": "for search request for Front End",
            "required": false
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Get Products",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/product"
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "require": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "model_number": {
                  "type": "string"
                },
                "main_image_url": {
                  "type": "string"
                },
                "price": {
                  "type": "string"
                },
                "size": {
                  "type": "string"
                },
                "material": {
                  "type": "string"
                },
                "catergory_id": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Create Product",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "insertedId": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/product/get-by-catergory-id/{catergory_id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "catergory_id",
            "type": "integer",
            "in": "path",
            "required": true
          },
          {
            "name": "productsPerPage",
            "in": "query",
            "description": "Max products can be display in one page",
            "required": false
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Page number",
            "required": false
          },
          {
            "name": "orderType",
            "in": "query",
            "description": "Order type to display: 1-oldest 2-newest ",
            "required": false
          },
          {
            "name": "search",
            "in": "query",
            "description": "for search request for Front End",
            "required": false
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Get Products By Catergory",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/product"
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/product/get-by-catergory-name/{name}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "type": "string",
            "in": "path",
            "description": "catergory_name",
            "required": true
          },
          {
            "name": "productsPerPage",
            "in": "query",
            "description": "Max products can be display in one page",
            "required": false
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Page number",
            "required": false
          },
          {
            "name": "orderType",
            "in": "query",
            "description": "Order type to display: 1-oldest 2-newest ",
            "required": false
          },
          {
            "name": "search",
            "in": "query",
            "description": "for search request for Front End",
            "required": false
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Get Products By Catergory",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/product"
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/product/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Get Product by Id",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/product"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "description": "ID of post",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "model_number": {
                  "type": "string"
                },
                "main_image_url": {
                  "type": "string"
                },
                "price": {
                  "type": "string"
                },
                "size": {
                  "type": "string"
                },
                "material": {
                  "type": "string"
                },
                "catergory_id": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Update Product by ID",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/success"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Delete product by ID",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/success"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/product/update-image/{product_id}": {
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "product_id",
            "type": "integer",
            "in": "path",
            "description": "ID of product",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "url_image1": {
                  "type": "string"
                },
                "url_image2": {
                  "type": "string"
                },
                "url_image3": {
                  "type": "string"
                },
                "url_image4": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Update Product Image by Product_ID",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/success"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/product/upload-image/{product_id}": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "product_id",
            "type": "integer",
            "in": "path",
            "description": "ID of product",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "url_image1": {
                  "type": "string"
                },
                "url_image2": {
                  "type": "string"
                },
                "url_image3": {
                  "type": "string"
                },
                "url_image4": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Updload Product Image by Product_ID",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/success"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "model_number": {
          "type": "string"
        },
        "main_image_url": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "size": {
          "type": "string"
        },
        "material": {
          "type": "string"
        },
        "catergory_id": {
          "$ref": "#/definitions/catergory"
        },
        "created_at": {
          "type": "string",
          "example": "2020-01-08T13:25:50"
        },
        "updated_at": {
          "type": "string",
          "example": "2020-01-08T13:25:50"
        }
      }
    },
...
    "main-catergory": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "image_url": {
          "type": "string"
        }
      }
    },
    "catergory": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "main_catergory": {
          "$ref": "#/definitions/main-catergory"
        }
      }
    },
...
    "error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "success": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}